# fixes #615 improve engine creation performance