@sebastienros 2019-05-17
I think it should be a huge gain for ravendb also. Will benchmark it on Orchard tomorrow, because we don't even use all these objects.