@lahma 2019-08-14
@sebastienros did you ever find the time to compare performance in Orchard?