@sebastienros 2019-08-14
I remember I checked and it was faster but don't recall the numbers.