@sebastienros 2019-08-14
if you like perf, there is so much to do in Orchard. It's allocating like crazy, which is not surprising because it does stuff ;) but which also means there are many ways to improve it.