@lahma 2019-08-14
Maybe at some point, probably requires quite a lot studying to understand hot paths and important scenarios...